requirements.txt
requests==2.32.3
pandas==2.3.2
pyyaml==6.0.2
tabulate==0.9.0
pytest==8.3.3
numpy<2
pyarrow==21.0.0